# Small dispatch fix for gamma