--- a/src/gamma.jl
+++ b/src/gamma.jl
@@ -1,6 +1,6 @@
 # Adapted from Cephes Mathematical Library (MIT license https://en.smath.com/view/CephesMathLibrary/license) by Stephen L. Moshier
 gamma(x::Float64) = _gamma(x)
-_gamma(x::Float32) = Float32(_gamma(Float64(x)))
+gamma(x::Float32) = Float32(_gamma(Float64(x)))
 
 function _gamma(x::Float64)
     T = Float64
@@ -52,7 +52,7 @@ end
 
 function gamma(n::Integer)
     n < 0 && throw(DomainError(n, "`n` must not be negative."))
-    n == 0 && return Inf*float(n)
+    n == 0 && return Inf*one(n)
     n > 20 && return gamma(float(n))
     @inbounds return Float64(factorial(n-1))
 end