~~Unfortunately we are promoting the argument to Float64 before hitting the gamma calls....~~

```julia
# This PR
julia> @benchmark Bessels.besselj(10, x) setup=(x=rand()*0.0000000001)
BenchmarkTools.Trial: 10000 samples with 996 evaluations.
 Range (min … max):  26.144 ns … 52.565 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     27.193 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   27.259 ns ±  0.996 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

                      ▂█▅                                      
  ▂▂▂▂▂▂▂▂▂▁▁▁▁▁▁▁▁▂▂▅███▆▃▄▃▂▁▁▁▁▁▁▁▁▁▁▁▁▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▂▁▂ ▂
  26.1 ns         Histogram: frequency by time          29 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.

# Master
julia> @benchmark Bessels.besselj(10, x) setup=(x=rand()*0.0000000001)
BenchmarkTools.Trial: 10000 samples with 988 evaluations.
 Range (min … max):  45.090 ns … 98.066 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     46.202 ns              ┊ GC (median):    0.00%
 Time  (mean ± σ):   46.266 ns ±  1.262 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

                    █▂  █▂                                     
  ▂▂▂▂▂▂▂▂▁▁▂▂▂▂▂▂▄███▄███▄▃▃▂▂▂▂▂▂▂▂▂▂▂▂▁▂▁▁▁▁▁▁▁▁▁▁▂▂▂▂▂▂▂▂ ▃
  45.1 ns         Histogram: frequency by time        48.2 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```
~~So this does improve computational time now..... but I'm not for sure we should do much more than this for #58.... hoping our Float32 tests are robust enough to catch if this promotion is even needed.~~

Edit: Just fixed the dispatch calls to actually hit the gamma integer branch and fixed the proper return value at x = 0. This does nothing to improve #58 as that promotion was needed for accuracy.