@heltonmc 2022-10-25
Ya that is not going to work we need that promotion for single precision. The result for gamma at zero returns NaN instead of Inf for 0 and -Inf for -0.0.